+++ backend/src/services/database/helpers.ts
@@ -1,5 +1,11 @@
 import { AppError } from '@/utils/errors.js';
-import { ERROR_CODES } from '@insforge/shared-schemas';
+import {
+  ERROR_CODES,
+  type ForeignKeySchema,
+  type OnDeleteActionSchema,
+  type OnUpdateActionSchema,
+} from '@insforge/shared-schemas';
+import type { ForeignKeyRow } from '@/types/database.js';
 import { validateIdentifier, validateSchemaName, validateTableName } from '@/utils/validations.js';
 
 export const DEFAULT_DATABASE_SCHEMA = 'public' as const;
@@ -74,4 +80,107 @@ export function splitQualifiedTableReference(
     schemaName,
     tableName,
   };
+}
+
+/**
+ * Parameterized pg_catalog query returning one row per foreign-key column pair,
+ * ordered by (constraint name, column ordinal position). A composite key produces
+ * one row per column; group the rows with {@link groupForeignKeyRows}.
+ *
+ * Params: $1 = schema name, $2 = table name.
+ *
+ * Single source of truth for FK introspection — used by both the table-schema
+ * service and the JSON export so the two can never drift (column ordering,
+ * referential-action mapping, schema qualification).
+ */
+export const FOREIGN_KEY_INTROSPECTION_QUERY = `
+  SELECT
+    c.conname AS constraint_name,
+    a1.attname AS from_column,
+    nf.nspname AS foreign_schema,
+    cf.relname AS foreign_table,
+    a2.attname AS foreign_column,
+    u.pos::int AS ordinal_position,
+    CASE c.confdeltype
+      WHEN 'a' THEN 'NO ACTION' WHEN 'r' THEN 'RESTRICT'
+      WHEN 'c' THEN 'CASCADE' WHEN 'n' THEN 'SET NULL'
+      WHEN 'd' THEN 'SET DEFAULT'
+    END AS on_delete,
+    CASE c.confupdtype
+      WHEN 'a' THEN 'NO ACTION' WHEN 'r' THEN 'RESTRICT'
+      WHEN 'c' THEN 'CASCADE' WHEN 'n' THEN 'SET NULL'
+      WHEN 'd' THEN 'SET DEFAULT'
+    END AS on_update
+  FROM pg_catalog.pg_constraint c
+  JOIN pg_catalog.pg_class ct ON c.conrelid = ct.oid
+  JOIN pg_catalog.pg_namespace nt ON ct.relnamespace = nt.oid
+  JOIN pg_catalog.pg_class cf ON c.confrelid = cf.oid
+  JOIN pg_catalog.pg_namespace nf ON cf.relnamespace = nf.oid
+  CROSS JOIN LATERAL unnest(c.conkey, c.confkey)
+    WITH ORDINALITY AS u(src_attnum, ref_attnum, pos)
+  JOIN pg_catalog.pg_attribute a1
+    ON a1.attnum = u.src_attnum AND a1.attrelid = c.conrelid
+  JOIN pg_catalog.pg_attribute a2
+    ON a2.attnum = u.ref_attnum AND a2.attrelid = c.confrelid
+  WHERE c.contype = 'f'
+    AND nt.nspname = $1
+    AND ct.relname = $2
+  ORDER BY c.conname, u.pos
+`;
+
+/**
+ * Groups the per-column rows from {@link FOREIGN_KEY_INTROSPECTION_QUERY} into one
+ * entity per constraint. A composite key becomes a single entity carrying all its
+ * (source -> reference) pairs in ordinal order. References to non-public schemas
+ * are qualified as `schema.table`.
+ */
+export function groupForeignKeyRows(rows: ForeignKeyRow[]): ForeignKeySchema[] {
+  const constraintGroups = new Map<
+    string,
+    {
+      constraintName: string;
+      referenceTable: string;
+      fromColumns: { name: string; foreignColumn: string; ordinal: number }[];
+      onDelete: string;
+      onUpdate: string;
+    }
+  >();
+
+  for (const fk of rows) {
+    const referenceTable =
+      fk.foreign_schema !== 'public'
+        ? `${fk.foreign_schema}.${fk.foreign_table}`
+        : fk.foreign_table;
+
+    let group = constraintGroups.get(fk.constraint_name);
+    if (!group) {
+      group = {
+        constraintName: fk.constraint_name,
+        referenceTable,
+        fromColumns: [],
+        onDelete: fk.on_delete,
+        onUpdate: fk.on_update,
+      };
+      constraintGroups.set(fk.constraint_name, group);
+    }
+    group.fromColumns.push({
+      name: fk.from_column,
+      foreignColumn: fk.foreign_column,
+      ordinal: fk.ordinal_position,
+    });
+  }
+
+  return Array.from(constraintGroups.values()).map((group) => {
+    group.fromColumns.sort((a, b) => a.ordinal - b.ordinal);
+    return {
+      constraintName: group.constraintName,
+      referenceTable: group.referenceTable,
+      referenceColumns: group.fromColumns.map((c) => ({
+        sourceColumn: c.name,
+        referenceColumn: c.foreignColumn,
+      })),
+      onDelete: group.onDelete as OnDeleteActionSchema,
+      onUpdate: group.onUpdate as OnUpdateActionSchema,
+    };
+  });
 }
